refactor(work): extract session lifecycle methods into session-lifecycle.ts
Move handleSessionEnd, finalizeTask, createPrFallback, buildIterationPrompt,
and cleanupWorktree (~300 lines) from WorkTaskService into a standalone
server/work/session-lifecycle.ts module. WorkTaskService retains thin wrappers
and a _lifecycleCtx() helper to pass dependencies.

service.ts: 1359 → 1175 lines; new session-lifecycle.ts: 234 lines.
